packages/fastapi/pyproject.toml
[tool.poetry]
name = "g2g"
version = "0.1.0"
description = "Get2Gether backend"
authors = ["Guro Bokum <jiojiajiu@gmail.com>"]
readme = "README.md"
packages = [{ include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.108.0"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
pydantic-settings = "^2.1.0"
asyncpg = "^0.29.0"
python-telegram-bot = "^21.1.1"
pyi18n-v2 = "^1.2.1"
pytest-asyncio = "^0.23.7"


[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"
ruff = "^0.1.11"
pytest = "^7.4.4"
pytest-watch = "^4.2.0"
taskipy = "^1.12.2"
coverage = "^7.4.0"
alembic = "^1.13.1"
pytest-dotenv = "^0.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# https://mypy.readthedocs.io/en/latest/config_file.html
[tool.mypy]
strict_optional = true
show_error_codes = true
show_column_numbers = true
disallow_any_unimported = true
implicit_reexport = false
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_no_return = true
warn_return_any = true
warn_unused_configs = true
strict_equality = true

[tool.pytest.ini_options]
env_files = ".env.test"
asyncio_mode = "auto"
env_override_existing_values = 1
filterwarnings = [
  "ignore:The 'app' shortcut is now deprecated. Use the explicit style 'transport=ASGITransport\\(app=...\\)':DeprecationWarning",
 ]

# https://docs.astral.sh/ruff/configuration
[tool.ruff]
line-length = 88 # default
extend-select = ["B", "I", "SIM", "UP"]

[tool.taskipy.tasks]
alembic_print = "bash alembic.sh"
makemigrations = "alembic revision --autogenerate -m"
migrate = "alembic upgrade head"
dev = "uvicorn app.server:app --reload"
lint = "ruff format --check . && ruff check --diff"
mypy = "mypy ."
test = "coverage run -m pytest && coverage report -m"
ptw = "ptw -v"
